requirements.txt
tensorflow>=2.12.0
streamlit>=1.22.0
numpy>=1.24.3
matplotlib>=3.7.1
Pillow>=9.5.0
streamlit-drawable-canvas